% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_CV.R
\name{cond_select}
\alias{cond_select}
\title{Neighborhood selection method with cross-validation to select the tuning parameter.}
\usage{
cond_select(
  data,
  formula,
  response,
  type = NULL,
  alpha,
  subset,
  na.action,
  rho,
  ydomain,
  yquad,
  prec,
  maxiter,
  skip.iter,
  M0,
  M_list,
  maxiteration,
  tolerance,
  id.basis = NULL,
  theta2,
  w2 = NULL,
  n,
  p
)
}
\arguments{
\item{data}{Data frame}

\item{formula}{Symbolic description of the model to be fit.}

\item{response}{Formula listing response variables.}

\item{type}{List specifying the type of spline for each variable.}

\item{alpha}{Parameter defining cross-validation score for smoothing parameter selection.}

\item{subset}{Optional vector specifying a subset of observations to be used in the fitting process.}

\item{na.action}{Function which indicates what should happen when the data contain NAs.}

\item{rho}{Method to construct rho function for neighborhood selection method.}

\item{ydomain}{Data frame specifying marginal support of conditional density in the neighborhood selection method.}

\item{yquad}{Quadrature for calculating integral on Y domain in the neighborhood selection method. Mandatory if response variables other than factors or numerical vectors are involved.}

\item{prec}{Precision requirement for internal iterations.}

\item{maxiter}{Maximum number of iterations allowed for internal iterations.}

\item{skip.iter}{Flag indicating whether to use initial values of theta and skip theta iteration in the neighborhood selection method.}

\item{M0}{Upper bound}

\item{M_list}{List of values for tuning parameter selection}

\item{maxiteration}{Max number of iteration}

\item{tolerance}{Threshold for convergence}

\item{id.basis}{Index of observations to be used as "knots."}

\item{theta2}{Parameters for edge selection.}

\item{w2}{Optional vector to specify weights for two-way interactions}

\item{n}{Number of observation}

\item{p}{Dimension of data frame}
}
\description{
Apply 5-fold cross-validation to select tuning parameter.
}
