% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gene.R
\docType{data}
\name{Gene}
\alias{Gene}
\title{Gene Data for Analysis}
\format{
An object of class \code{data.frame} with 118 rows and 39 columns.
}
\source{
Anja Wille, Philip Zimmermann, Eva Vranova, Andreas Furholz, Oliver Laule, Stefan
Bleuler, Lars Hennig, Amela Prelíc, Peter von Rohr, Lothar Thiele, et al. Sparse grahpical
 gaussian modeling of the isoprenoid gene network in arabidopsis thaliana. Genome
biology, 5(11):1–13, 2004
}
\usage{
Gene
}
\description{
There are 118 observations of Affymetrix GeneChip microarrays in the dataset,
where a subset of 39 genes from the isoprenoid pathway is selected for analysis.
}
\examples{
data(Gene)
}
\keyword{datasets}
