% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint.R
\name{selection.joint}
\alias{selection.joint}
\title{Joint method for edge selection}
\usage{
selection.joint(data, ...)
}
\arguments{
\item{data}{Data frame}

\item{...}{Any options can be defined.
\itemize{
\item \code{type} List specifying the type of spline for each variable.
\item \code{alpha} Parameter defining cross-validation score for smoothing parameter selection.
\item \code{subset} Optional vector specifying a subset of observations to be used in the fitting process.
\item \code{na.action} Function which indicates what should happen when the data contain NAs.
\item \code{seed} Seed to be used for the random generation of "knots."
\item \code{prec} Precision requirement for internal iterations.
\item \code{maxiter} Maximum number of iterations allowed for internal iterations.
\item \code{id.basis} Index of observations to be used as "knots."
\item \code{nbasis} Number of "knots" to be used.
\item \code{domain} Data frame specifying marginal support of density in the joint method.
\item \code{quad} Quadrature for calculating integral in the joint method. Mandatory if variables other than factors or numerical vectors are involved.
\item \code{w} Optional vector to specify weights for two-way interactions in the joint method.
}}
}
\description{
Apply joint selection method by considering joint distribution of all variables.
}
\examples{
library(gss)
data(NO2)
edge.selection(data = NO2, family = "joint", nbasis = 100)
}
