% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semi.R
\name{selection.semi}
\alias{selection.semi}
\title{Semi-parametric method for edge selection.}
\usage{
selection.semi(data, ...)
}
\arguments{
\item{data}{Data Frame.}

\item{...}{Any options can be defined.
\itemize{
\item \code{px} Dimension of variables in the semi-parametric method to be estimated using non-parametric method.
\item \code{maxLambda} Number used to generate the range of the tuning parameter for selection of Lambda matrix.
\item \code{maxTheta} Number used to generate the range of the tuning parameter for selection of Theta matrix.
\item \code{iterLambda} Number of iterations to find optimal Lambda matrix.
\item \code{iterTheta} Number of iterations to find optimal Theta matrix.
\item \code{cutoff} Cutoff value for squared projection.
\item \code{N} Number of simulations in non-parametric part to calculate the standard error.
\item \code{semiMethod} Method type to select optimal matrix in semi-parametric method.
}}
}
\description{
Semi-parametric method for edge selection.
}
\examples{
library(huge)
n <- 200; p <- 20; px <- 5
# Simulate high dimension data
set.seed(5732)
z <- huge.generator(n, d = p, graph = "random", prob = .2, verbose = FALSE, vis = FALSE, v = .65)
data <- data.frame(z$data)
# Assume we estimate the first five variables using non-parametric method.
edge.selection(data = data, family ="semi", px = px)
}
