% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sscden0.R
\name{sscden0}
\alias{sscden0}
\title{Fit conditional density model}
\usage{
sscden0(
  formula,
  response,
  type = NULL,
  data = list(),
  weights,
  subset,
  na.action = na.omit,
  alpha = 1.4,
  id.basis = NULL,
  nbasis = NULL,
  seed = NULL,
  ydomain = as.list(NULL),
  yquad = NULL,
  prec = 1e-06,
  maxiter = 30,
  skip.iter = FALSE
)
}
\arguments{
\item{formula}{Symbolic description of the model to be fit.}

\item{response}{Formula listing response variables.}

\item{type}{List specifying the type of spline for each variable.}

\item{data}{Optional data frame containing the variables in the model.}

\item{weights}{Optional vector of bin-counts for histogram data.}

\item{subset}{Optional vector specifying a subset of observations to be used in the fitting process.}

\item{na.action}{Function which indicates what should happen when the data contain NAs.}

\item{alpha}{Parameter defining cross-validation score for smoothing parameter selection.}

\item{id.basis}{Index of observations to be used as "knots."}

\item{nbasis}{Number of "knots" to be used.}

\item{seed}{Seed to be used for the random generation of "knots."}

\item{ydomain}{Data frame specifying marginal support of conditional density.}

\item{yquad}{Quadrature for calculating integral on Y domain. Mandatory if response variables other than factors or numerical vectors are involved.}

\item{prec}{Precision requirement for internal iterations.}

\item{maxiter}{Maximum number of iterations allowed for internal iterations.}

\item{skip.iter}{Flag indicating whether to use initial values of theta and skip theta iteration.}
}
\description{
Fit conditional density model to initialize parameters.
}
