% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sscden_selection.R
\name{sscden_selection}
\alias{sscden_selection}
\title{Fit conditional density model}
\usage{
sscden_selection(
  formula,
  response,
  type = NULL,
  data = list(),
  weights,
  subset,
  na.action = na.omit,
  alpha = 1.4,
  id.basis = NULL,
  nbasis = NULL,
  seed = NULL,
  rho = list("xy"),
  ydomain = as.list(NULL),
  yquad = NULL,
  prec = 1e-07,
  maxiter = 30,
  skip.iter = TRUE,
  p = 2,
  theta2 = NULL,
  w2 = NULL
)
}
\arguments{
\item{formula}{Symbolic description of the model to be fit.}

\item{response}{Formula listing response variables.}

\item{type}{List specifying the type of spline for each variable.}

\item{data}{Optional data frame containing the variables in the model.}

\item{weights}{Optional vector of bin-counts for histogram data.}

\item{subset}{Optional vector specifying a subset of observations to be used in the fitting process.}

\item{na.action}{Function which indicates what should happen when the data contain NAs.}

\item{alpha}{Parameter defining cross-validation score for smoothing parameter selection.}

\item{id.basis}{Index of observations to be used as "knots."}

\item{nbasis}{Number of "knots" to be used.}

\item{seed}{Seed to be used for the random generation of "knots."}

\item{rho}{Method to construct rho function.}

\item{ydomain}{Data frame specifying marginal support of conditional density.}

\item{yquad}{Quadrature for calculating integral on Y domain. Mandatory if response variables other than factors or numerical vectors are involved.}

\item{prec}{Precision requirement for internal iterations.}

\item{maxiter}{Maximum number of iterations allowed for internal iterations.}

\item{skip.iter}{Flag indicating whether to use initial values of theta and skip theta iteration.}

\item{p}{Dimension of data frame.}

\item{theta2}{Parameters for two-way interactions.}

\item{w2}{Weights for two-way interactions.}
}
\description{
Fit conditional density model
}
