% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssden0.R
\name{ssden0}
\alias{ssden0}
\title{Generate quadrature for conditional density fit}
\usage{
ssden0(
  formula,
  type = NULL,
  data = list(),
  alpha = 1.4,
  weights = NULL,
  subset,
  na.action = na.omit,
  id.basis = NULL,
  nbasis = NULL,
  seed = NULL,
  domain = as.list(NULL),
  quad = NULL,
  qdsz.depth = NULL,
  bias = NULL,
  prec = 1e-07,
  maxiter = 30,
  skip.iter = TRUE
)
}
\arguments{
\item{formula}{Symbolic description of the model to be fit.}

\item{type}{List specifying the type of spline for each variable.}

\item{data}{Optional data frame containing the variables in the model.}

\item{alpha}{Parameter defining cross-validation score for smoothing parameter selection.}

\item{weights}{Optional vector of bin-counts for histogram data.}

\item{subset}{Optional vector specifying a subset of observations to be used in the fitting process.}

\item{na.action}{Function which indicates what should happen when the data contain NAs.}

\item{id.basis}{Index of observations to be used as "knots."}

\item{nbasis}{Number of "knots" to be used.}

\item{seed}{Seed to be used for the random generation of "knots."}

\item{domain}{Data frame specifying marginal support of density.}

\item{quad}{Quadrature for calculating integral. Mandatory if variables other than factors or numerical vectors are involved.}

\item{qdsz.depth}{Depth for the generation of quadrature.}

\item{bias}{Input for sampling bias.}

\item{prec}{Precision requirement for internal iterations.}

\item{maxiter}{Maximum number of iterations allowed for internal iterations.}

\item{skip.iter}{Flag indicating whether to use initial values of theta and skip theta iteration.}
}
\description{
Generate quadrature for conditional density fit
}
