% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssden2.R
\name{ssden2}
\alias{ssden2}
\title{Estimating joint density using smoothing splines}
\usage{
ssden2(
  formula,
  type = NULL,
  data = list(),
  alpha = 1.4,
  weights = NULL,
  subset,
  na.action = na.omit,
  id.basis = NULL,
  nbasis = NULL,
  seed = NULL,
  domain = as.list(NULL),
  quad = NULL,
  prec = 1e-07,
  maxiter = 30,
  theta2 = NULL,
  w = NULL
)
}
\arguments{
\item{formula}{Symbolic description of the model to be fit.}

\item{type}{List specifying the type of spline for each variable.}

\item{data}{Optional data frame containing the variables in the model.}

\item{alpha}{Parameter defining cross-validation score for smoothing parameter selection.}

\item{weights}{Optional vector of bin-counts for histogram data.}

\item{subset}{Optional vector specifying a subset of observations to be used in the fitting process.}

\item{na.action}{Function which indicates what should happen when the data contain NAs.}

\item{id.basis}{Index of observations to be used as "knots."}

\item{nbasis}{Number of "knots" to be used.}

\item{seed}{Seed to be used for the random generation of "knots."}

\item{domain}{Data frame specifying marginal support of density.}

\item{quad}{Quadrature for calculating integral. Mandatory if variables other than factors or numerical vectors are involved.}

\item{prec}{Precision requirement for internal iterations.}

\item{maxiter}{Maximum number of iterations allowed for internal iterations.}

\item{theta2}{Parameters for two-way interactions.}

\item{w}{Optional vector to specify weights for two-way interactions.}
}
\description{
Estimating joint density using smoothing splines
}
